test(mutation-kill): split oversized test files by scenario
tests/scripts/test_mutation_kill_loop.py and tests/agents/test_mutation_kill_agent.py
grew past the 500-line guideline. Split by responsibility: config/scoped-run mechanics
stay in the original files, orchestration/insertion/verification/CLI move to new
sibling files, and feasibility-gate doc tests move to their own file mirroring the
existing baseline-reuse split. Shared fixtures centralized into
_mutation_kill_loop_test_helpers.py instead of duplicated per file.

Light review (test-review, test-smell-review, structure-review) caught a real
defect this split introduced: test_mutation_kill_loop_cli.py only resolved its
script imports by accident of pytest collection order (relying on a sibling
file's sys.path side effect), failing when run in isolation. Fixed by importing
_mutation_kill_loop_test_helpers explicitly for its sys.path side effect, and
corrected two related docstring misattributions of which file owns which tests.

No behavior change otherwise: 147/147 tests preserved across the split.

Closes #1564
