+++ tests/scripts/test_mutation_kill_loop_verify.py
@@ -1,0 +1,497 @@
+"""Pytest tests for mutation_kill_loop.py's verify/commit subprocess wiring —
+``dotnet_build``/``dotnet_test``/``git_revert``/``git_commit``/
+``git_reset_and_revert`` (#1564 split of ``test_mutation_kill_loop.py``).
+Every dotnet/git subprocess is mocked; no real .NET tooling runs.
+
+The script's CLI dispatch and the cross-module no-repo-specific-literal sweep
+live in ``test_mutation_kill_loop_cli.py`` — a separate file so this one stays
+scoped to the verify/commit subprocess primitives alone.
+"""
+
+from __future__ import annotations
+
+import pytest
+from _mutation_kill_loop_test_helpers import _write_config
+
+import mutation_kill_loop as loop  # noqa: E402
+
+
+# =============================================================================
+# Verify/commit subprocess wiring — targets come from config (no literals)
+# =============================================================================
+def test_dotnet_build_uses_configured_test_project(
+    tmp_path, monkeypatch: pytest.MonkeyPatch
+):
+    config = loop.load_loop_config(
+        _write_config(tmp_path, test_projects=("test/Foo.Tests/Foo.Tests.csproj",))
+    )
+    seen: list = []
+
+    class _R:
+        returncode = 0
+        stdout = ""
+        stderr = ""
+
+    monkeypatch.setattr(
+        loop.subprocess, "run", lambda argv, **k: seen.append(argv) or _R()
+    )
+
+    assert loop.dotnet_build(loop.dotnet_test_project_targets(config)) is True
+    assert seen[0] == [
+        "dotnet",
+        "build",
+        "test/Foo.Tests/Foo.Tests.csproj",
+        "-c",
+        "Debug",
+        "--nologo",
+    ]
+
+
+# =============================================================================
+# Scenario: A hung `dotnet build`/`dotnet test` is bounded by a timeout — it
+# fails (and reverts) rather than hanging the loop forever (#1558)
+# =============================================================================
+def test_dotnet_build_passes_a_timeout(monkeypatch: pytest.MonkeyPatch):
+    captured: dict = {}
+
+    def fake_run(argv, **kwargs):
+        captured.update(kwargs)
+
+        class _R:
+            returncode = 0
+
+        return _R()
+
+    monkeypatch.setattr(loop.subprocess, "run", fake_run)
+
+    loop.dotnet_build(["test/Foo.Tests/Foo.Tests.csproj"])
+
+    assert captured["timeout"] == loop.DOTNET_BUILD_TIMEOUT_S
+
+
+def test_dotnet_build_timeout_is_treated_as_a_build_failure(
+    monkeypatch: pytest.MonkeyPatch, capsys
+):
+    def fake_run(argv, **kwargs):
+        raise loop.subprocess.TimeoutExpired(argv, kwargs["timeout"])
+
+    monkeypatch.setattr(loop.subprocess, "run", fake_run)
+
+    assert loop.dotnet_build(["test/Foo.Tests/Foo.Tests.csproj"]) is False
+    err = capsys.readouterr().err
+    assert str(loop.DOTNET_BUILD_TIMEOUT_S) in err
+    assert "DEV_TEAM_MUTATION_BUILD_TIMEOUT_S" in err
+
+
+def test_dotnet_build_returns_false_on_nonzero_returncode(
+    monkeypatch: pytest.MonkeyPatch
+):
+    """dotnet_build's own failure-path logic (a non-zero exit, no timeout
+    involved) — direct coverage, not mocked away at the orchestration level
+    (#1584)."""
+
+    class _R:
+        returncode = 1
+        stdout = ""
+        stderr = "error CS0000: something broke\n"
+
+    monkeypatch.setattr(loop.subprocess, "run", lambda argv, **k: _R())
+
+    assert loop.dotnet_build(["Foo.Tests.csproj"]) is False
+
+
+def test_dotnet_test_returns_false_on_nonzero_returncode_even_with_zero_failed(
+    monkeypatch: pytest.MonkeyPatch
+):
+    """A non-zero `dotnet test` exit fails the round even when the "Failed:"
+    line itself parses to 0 (e.g. the process crashed before reporting) —
+    direct coverage of dotnet_test's own failure-path logic (#1584)."""
+
+    class _R:
+        returncode = 1
+        stdout = "Failed: 0, Passed: 5\n"
+        stderr = ""
+
+    monkeypatch.setattr(loop.subprocess, "run", lambda argv, **k: _R())
+
+    assert loop.dotnet_test(["Foo.Tests.csproj"], "FooTests") is False
+
+
+def test_dotnet_test_accumulates_failed_count_across_multiple_assemblies(
+    monkeypatch: pytest.MonkeyPatch
+):
+    """A single `dotnet test` invocation spanning multiple test assemblies
+    prints one "Failed: N" line per assembly. The scan must SUM them, not
+    let a later, clean assembly's "Failed: 0" line overwrite an earlier
+    assembly's real failures — a last-match-wins scan would falsely report
+    success here (#1598). Fixture shape matches real `dotnet test` per-
+    assembly summary lines (item 11, #1598/#1584 review), not a synthetic
+    shorthand — see also the rollup-line non-double-count test below."""
+
+    class _R:
+        returncode = 0
+        stdout = (
+            "Failed!  - Failed:     3, Passed:     7, Skipped:     0, "
+            "Total:    10, Duration: 120 ms - Bar.Tests.dll\n"
+            "Passed!  - Failed:     0, Passed:    12, Skipped:     0, "
+            "Total:    12, Duration: 40 ms - Foo.Tests.dll\n"
+        )
+        stderr = ""
+
+    monkeypatch.setattr(loop.subprocess, "run", lambda argv, **k: _R())
+
+    assert loop.dotnet_test(["Foo.Tests.csproj"], "FooTests") is False
+
+
+def test_dotnet_test_trailing_lowercase_rollup_line_does_not_flip_a_clean_run(
+    monkeypatch: pytest.MonkeyPatch
+):
+    """Newer `dotnet test` SDKs (7/8) also print a final, lowercase rollup
+    line ("Test summary: total: ... failed: 0 ...") after the per-assembly
+    summaries. The scan is capital-F "Failed:" only, so this rollup line
+    must not be mistaken for a second, independent "Failed:" occurrence —
+    a fully clean multi-assembly run (returncode 0, every real "Failed:"
+    count is 0) must still report success (item 11, #1598/#1584 review)."""
+
+    class _R:
+        returncode = 0
+        stdout = (
+            "Passed!  - Failed:     0, Passed:    10, Skipped:     0, "
+            "Total:    10, Duration: 40 ms - Foo.Tests.dll\n"
+            "\n"
+            "Test summary: total: 10, failed: 0, succeeded: 10, skipped: 0, "
+            "duration: 0.1s\n"
+        )
+        stderr = ""
+
+    monkeypatch.setattr(loop.subprocess, "run", lambda argv, **k: _R())
+
+    assert loop.dotnet_test(["Foo.Tests.csproj"], "FooTests") is True
+
+
+def test_sum_failed_counts_is_case_sensitive_not_double_counting_the_rollup_line():
+    """A fixture where both a correct (case-sensitive) scan and a buggy
+    (case-insensitive) scan sum to 0 can't actually catch a regression to
+    ``re.IGNORECASE`` (#1598/#1584 review, item 12) — the boolean
+    ``dotnet_test`` result would be identical either way. Asserting the
+    NUMERIC total directly against a nonzero per-assembly count (3) plus a
+    nonzero lowercase rollup count (3) proves the scan is case-sensitive:
+    a case-insensitive regex would sum both `Failed:`-shaped occurrences
+    (3 + 3 = 6); the correct, case-sensitive scan counts only the real
+    capital-F per-assembly line (3)."""
+    text = (
+        "Failed!  - Failed:     3, Passed:     7, Skipped:     0, "
+        "Total:    10, Duration: 120 ms - Bar.Tests.dll\n"
+        "\n"
+        "Test summary: total: 10, failed: 3, succeeded: 7, skipped: 0, "
+        "duration: 0.1s\n"
+    )
+
+    assert loop._sum_failed_counts(text) == 3
+
+
+def test_dotnet_test_passes_a_timeout(monkeypatch: pytest.MonkeyPatch):
+    captured: dict = {}
+
+    class _R:
+        returncode = 0
+        stdout = ""
+        stderr = ""
+
+    def fake_run(argv, **kwargs):
+        captured.update(kwargs)
+        return _R()
+
+    monkeypatch.setattr(loop.subprocess, "run", fake_run)
+
+    assert loop.dotnet_test(["test/Foo.Tests/Foo.Tests.csproj"], "FooTests") is True
+    assert captured["timeout"] == loop.DOTNET_TEST_TIMEOUT_S
+
+
+def test_dotnet_test_timeout_is_treated_as_a_test_failure(
+    monkeypatch: pytest.MonkeyPatch, capsys
+):
+    def fake_run(argv, **kwargs):
+        raise loop.subprocess.TimeoutExpired(argv, kwargs["timeout"])
+
+    monkeypatch.setattr(loop.subprocess, "run", fake_run)
+
+    assert loop.dotnet_test(["test/Foo.Tests/Foo.Tests.csproj"], "FooTests") is False
+    err = capsys.readouterr().err
+    assert str(loop.DOTNET_TEST_TIMEOUT_S) in err
+    assert "DEV_TEAM_MUTATION_TEST_TIMEOUT_S" in err
+
+
+def test_git_revert_checks_out_only_the_test_file(
+    tmp_path, monkeypatch: pytest.MonkeyPatch
+):
+    seen: list = []
+    monkeypatch.setattr(loop.subprocess, "run", lambda argv, **k: seen.append(argv))
+
+    loop.git_revert(tmp_path / "PaymentServiceTests.cs")
+
+    assert seen[0] == [
+        "git",
+        "--literal-pathspecs",
+        "checkout",
+        "--",
+        str(tmp_path / "PaymentServiceTests.cs"),
+    ]
+
+
+# =============================================================================
+# Scenario: A hung `git checkout`/`git add`/`git commit` is bounded by a
+# timeout, not left to hang the loop forever (#1572 — same class of fix as
+# #1558's dotnet_build/dotnet_test/Stryker timeouts).
+# =============================================================================
+def test_git_revert_passes_a_timeout(tmp_path, monkeypatch: pytest.MonkeyPatch):
+    captured: dict = {}
+    monkeypatch.setattr(loop.subprocess, "run", lambda argv, **k: captured.update(k))
+
+    loop.git_revert(tmp_path / "PaymentServiceTests.cs")
+
+    assert captured["timeout"] == loop.GIT_TIMEOUT_S
+
+
+def test_git_revert_timeout_is_logged_not_raised(
+    tmp_path, monkeypatch: pytest.MonkeyPatch, capsys
+):
+    def fake_run(argv, **kwargs):
+        raise loop.subprocess.TimeoutExpired(argv, kwargs["timeout"])
+
+    monkeypatch.setattr(loop.subprocess, "run", fake_run)
+
+    loop.git_revert(tmp_path / "PaymentServiceTests.cs")  # must not raise
+
+    err = capsys.readouterr().err
+    assert str(loop.GIT_TIMEOUT_S) in err
+    assert "DEV_TEAM_MUTATION_GIT_TIMEOUT_S" in err
+    # git_revert, git_commit's "add" leg, and git_commit's "commit" leg all
+    # share GIT_TIMEOUT_S/DEV_TEAM_MUTATION_GIT_TIMEOUT_S — the operation_label text
+    # is what distinguishes their timeout messages from one another.
+    assert "git checkout" in err
+
+
+def test_git_commit_passes_a_timeout_to_add_and_commit(
+    tmp_path, monkeypatch: pytest.MonkeyPatch
+):
+    calls: list = []
+
+    class _R:
+        returncode = 0
+
+    def fake_run(argv, **kwargs):
+        calls.append((argv, kwargs))
+        return _R()
+
+    monkeypatch.setattr(loop.subprocess, "run", fake_run)
+
+    result = loop.git_commit("msg", tmp_path / "PaymentServiceTests.cs")
+
+    assert result is True
+    assert calls[0][0] == [
+        "git",
+        "--literal-pathspecs",
+        "add",
+        "--",
+        str(tmp_path / "PaymentServiceTests.cs"),
+    ]
+    assert calls[0][1]["timeout"] == loop.GIT_TIMEOUT_S
+    assert calls[1][0][:3] == ["git", "--literal-pathspecs", "commit"]
+    assert calls[1][1]["timeout"] == loop.GIT_TIMEOUT_S
+
+
+def test_git_commit_add_leg_timeout_returns_false(
+    tmp_path, monkeypatch: pytest.MonkeyPatch, capsys
+):
+    def fake_run(argv, **kwargs):
+        raise loop.subprocess.TimeoutExpired(argv, kwargs["timeout"])
+
+    monkeypatch.setattr(loop.subprocess, "run", fake_run)
+
+    assert loop.git_commit("msg", tmp_path / "PaymentServiceTests.cs") is False
+    err = capsys.readouterr().err
+    assert str(loop.GIT_TIMEOUT_S) in err
+    assert "DEV_TEAM_MUTATION_GIT_TIMEOUT_S" in err
+    assert "git add" in err
+
+
+def test_git_commit_commit_leg_timeout_returns_false(
+    tmp_path, monkeypatch: pytest.MonkeyPatch, capsys
+):
+    """The `git add` leg succeeds; only the `git commit` leg's own
+    _run_with_timeout call times out — a distinct branch from the add-leg
+    timeout above, since git_commit short-circuits on add before ever
+    reaching commit."""
+
+    calls: list = []
+
+    class _R:
+        returncode = 0
+
+    def fake_run(argv, **kwargs):
+        calls.append(argv)
+        if len(calls) == 1:
+            return _R()
+        raise loop.subprocess.TimeoutExpired(argv, kwargs["timeout"])
+
+    monkeypatch.setattr(loop.subprocess, "run", fake_run)
+
+    assert loop.git_commit("msg", tmp_path / "PaymentServiceTests.cs") is False
+    # Pin the branch this test targets by call order, not just count — a
+    # future reordering of the add/commit calls should fail loudly here
+    # rather than silently exercising the wrong leg.
+    assert calls[0][:3] == ["git", "--literal-pathspecs", "add"]
+    err = capsys.readouterr().err
+    assert str(loop.GIT_TIMEOUT_S) in err
+    assert "DEV_TEAM_MUTATION_GIT_TIMEOUT_S" in err
+    assert "git commit" in err
+
+
+def test_git_commit_scopes_the_commit_call_to_the_test_file(
+    tmp_path, monkeypatch: pytest.MonkeyPatch
+):
+    """git commit -m message with no pathspec commits the WHOLE index, not
+    just test_file — the commit call must carry the same `-- <path>`
+    pathspec `git add` already uses (#1598)."""
+    calls: list = []
+
+    class _R:
+        returncode = 0
+
+    def fake_run(argv, **kwargs):
+        calls.append(argv)
+        return _R()
+
+    monkeypatch.setattr(loop.subprocess, "run", fake_run)
+
+    result = loop.git_commit("msg", tmp_path / "PaymentServiceTests.cs")
+
+    assert result is True
+    assert calls[1] == [
+        "git",
+        "--literal-pathspecs",
+        "commit",
+        "-m",
+        "msg",
+        "--",
+        str(tmp_path / "PaymentServiceTests.cs"),
+    ]
+
+
+def test_git_revert_returns_true_on_success(
+    tmp_path, monkeypatch: pytest.MonkeyPatch
+):
+    class _R:
+        returncode = 0
+
+    monkeypatch.setattr(loop.subprocess, "run", lambda argv, **k: _R())
+
+    assert loop.git_revert(tmp_path / "PaymentServiceTests.cs") is True
+
+
+def test_git_revert_returns_false_on_nonzero_exit(
+    tmp_path, monkeypatch: pytest.MonkeyPatch
+):
+    class _R:
+        returncode = 1
+
+    monkeypatch.setattr(loop.subprocess, "run", lambda argv, **k: _R())
+
+    assert loop.git_revert(tmp_path / "PaymentServiceTests.cs") is False
+
+
+def test_git_revert_returns_false_on_timeout(
+    tmp_path, monkeypatch: pytest.MonkeyPatch
+):
+    def fake_run(argv, **kwargs):
+        raise loop.subprocess.TimeoutExpired(argv, kwargs["timeout"])
+
+    monkeypatch.setattr(loop.subprocess, "run", fake_run)
+
+    assert loop.git_revert(tmp_path / "PaymentServiceTests.cs") is False
+
+
+def test_git_commit_proceeds_to_commit_after_a_non_timeout_add_failure(
+    tmp_path, monkeypatch: pytest.MonkeyPatch
+):
+    """Pins the documented (not-changed-by-#1593) behavior: git_commit only
+    inspects git add's timeout, not its returncode, so a non-timeout git add
+    failure (e.g. bad pathspec) still falls through to attempting the
+    commit."""
+
+    calls: list = []
+
+    class _R:
+        def __init__(self, returncode):
+            self.returncode = returncode
+
+    def fake_run(argv, **kwargs):
+        calls.append(argv)
+        return _R(returncode=1) if len(calls) == 1 else _R(returncode=0)
+
+    monkeypatch.setattr(loop.subprocess, "run", fake_run)
+
+    result = loop.git_commit("msg", tmp_path / "PaymentServiceTests.cs")
+
+    assert result is True
+    assert len(calls) == 2
+    assert calls[1][:3] == ["git", "--literal-pathspecs", "commit"]
+
+
+# =============================================================================
+# Scenario: git_reset_and_revert's own internal failure branches — every
+# other test either exercises the full real-git success path or mocks
+# git_reset_and_revert away wholesale (#1598/#1584 review, item 10).
+# =============================================================================
+def test_git_reset_and_revert_returns_false_when_reset_fails(
+    tmp_path, monkeypatch: pytest.MonkeyPatch
+):
+    calls: list = []
+
+    class _R:
+        def __init__(self, returncode):
+            self.returncode = returncode
+
+    def fake_run(argv, **kwargs):
+        calls.append(argv)
+        return _R(1)  # reset fails
+
+    monkeypatch.setattr(loop.subprocess, "run", fake_run)
+
+    assert loop.git_reset_and_revert(tmp_path / "PaymentServiceTests.cs") is False
+    # Only the reset call happens — checkout (git_revert) is never reached.
+    assert len(calls) == 1
+    assert calls[0][:3] == ["git", "--literal-pathspecs", "reset"]
+
+
+def test_git_reset_and_revert_returns_false_when_checkout_fails_after_reset_succeeds(
+    tmp_path, monkeypatch: pytest.MonkeyPatch
+):
+    calls: list = []
+
+    class _R:
+        def __init__(self, returncode):
+            self.returncode = returncode
+
+    def fake_run(argv, **kwargs):
+        calls.append(argv)
+        if argv[2] == "reset":
+            return _R(0)  # reset succeeds
+        return _R(1)  # checkout (git_revert) fails
+
+    monkeypatch.setattr(loop.subprocess, "run", fake_run)
+
+    assert loop.git_reset_and_revert(tmp_path / "PaymentServiceTests.cs") is False
+    assert [c[2] for c in calls] == ["reset", "checkout"]
+
+
+def test_git_reset_and_revert_returns_true_when_both_steps_succeed(
+    tmp_path, monkeypatch: pytest.MonkeyPatch
+):
+    class _R:
+        returncode = 0
+
+    monkeypatch.setattr(loop.subprocess, "run", lambda argv, **k: _R())
+
+    assert loop.git_reset_and_revert(tmp_path / "PaymentServiceTests.cs") is True
